+++ src/plugins/select/alps/select_alps.c
@@ -357,10 +357,10 @@ extern int select_p_job_ready(struct job_record *job_ptr)
 	 *		means that we need to confirm only if batch_flag is 0,
 	 *		and execute the other_job_ready() only in slurmctld.
 	 */
-	if ((slurmctld_primary || (job_ptr->job_state == (uint16_t)NO_VAL))
+	if ((slurmctld_primary || (job_ptr->job_state == NO_VAL))
 	    && !job_ptr->batch_flag && !_zero_size_job(job_ptr))
 		rc = do_basil_confirm(job_ptr);
-	if ((rc != SLURM_SUCCESS) || (job_ptr->job_state == (uint16_t) NO_VAL))
+	if ((rc != SLURM_SUCCESS) || (job_ptr->job_state == NO_VAL))
 		return rc;
 	return other_job_ready(job_ptr);
 }
@@ -444,7 +444,7 @@ extern int select_p_job_fini(struct job_record *job_ptr)
 	 * handled on the stepd end.
 	 */
 	if (((slurmctld_primary && !job_ptr->batch_flag) ||
-	     (job_ptr->job_state == (uint16_t)NO_VAL))
+	     (job_ptr->job_state == NO_VAL))
 	    && !_zero_size_job(job_ptr) &&
 	    (do_basil_release(job_ptr) != SLURM_SUCCESS))
 		return SLURM_ERROR;
@@ -453,7 +453,7 @@ extern int select_p_job_fini(struct job_record *job_ptr)
 	 *             stepdmgr, where job_state == NO_VAL is used to
 	 *             distinguish the context from that of slurmctld.
 	 */
-	if (job_ptr->job_state == (uint16_t)NO_VAL)
+	if (job_ptr->job_state == NO_VAL)
 		return SLURM_SUCCESS;
 	return other_job_fini(job_ptr);
 }
